+++ src/main/kotlin/com/github/diekautz/ideplugin/utils/IOUtilities.kt
@@ -15,8 +15,6 @@ import com.intellij.openapi.application.runReadAction
 import com.intellij.openapi.application.runWriteAction
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.diagnostic.thisLogger
-import com.intellij.openapi.fileChooser.FileChooserFactory
-import com.intellij.openapi.fileChooser.FileSaverDescriptor
 import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.options.ShowSettingsUtil
 import com.intellij.openapi.project.Project
@@ -110,16 +108,6 @@ fun screenshotFilesInEditor(project: Project, filePaths: List<String>): Map<Stri
             return@associateWith null
         }
         editor.screenshot()
-    }
-}
-
-inline fun <reified T> askAndSaveToDisk(project: Project, data: T, dialogTitle: String, filename: String? = null) {
-    val descriptor = FileSaverDescriptor(dialogTitle, "", ".json")
-    val saveFileDialog = FileChooserFactory.getInstance().createSaveFileDialog(descriptor, project)
-
-    val file = saveFileDialog.save(filename)?.file
-    if (file != null) {
-        saveToDisk(data, file)
     }
 }